+++ src/lib/cli.ts
@@ -95,7 +95,8 @@ async function run(app: td.Application) {
 
     const preValidationWarnCount = app.logger.warningCount;
     app.validate(project);
-    const hadValidationWarnings = app.logger.warningCount !== preValidationWarnCount;
+    const hadValidationWarnings = app.logger.warningCount !== preValidationWarnCount ||
+        app.logger.validationWarningCount != 0;
     if (app.logger.hasErrors()) {
         return ExitCodes.ValidationError;
     }